requirements.txt
# MCP Server Dependencies
mcp>=0.9.0

# Google API Dependencies
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Environment Variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0